New file structure with smaller files
