+++ src/lib/classes/automaton/backend.ts
@@ -1,0 +1,6 @@
+
+export enum Backend {
+	J_ECDAR = 0,
+	REVEAAL = 1
+}
+
