+++ src/lib/classes/automaton/location_type.ts
@@ -1,0 +1,9 @@
+
+export enum LocationType {
+	INITIAL = 'INITIAL',
+	NORMAL = 'NORMAL',
+	UNIVERSAL = 'UNIVERSAL',
+	INCONSISTENT = 'INCONSISTENT',
+	ANY = 'ANY'
+}
+
