+++ src/lib/classes/automaton/operator_type.ts
@@ -1,0 +1,8 @@
+
+export enum OperatorType {
+	COMPOSITION = 'COMPOSITION',
+	CONJUNCTION = 'CONJUNCTION',
+	REFINEMENT = 'REFINEMENT',
+	QUOTIENT = 'QUOTIENT',
+	SIMPLE = 'SIMPLE'
+}
