+++ src/lib/classes/automaton/raw/declaration.ts
@@ -1,0 +1,9 @@
+
+/**
+* The raw Object for a Declaration that is used to save and communicate in JSON.
+* */
+export type RawDeclaration = {
+	name: string;
+	declarations: string;
+};
+
