+++ src/lib/classes/automaton/status.ts
@@ -1,0 +1,6 @@
+
+export enum Status {
+	INPUT = 'INPUT',
+	OUTPUT = 'OUTPUT'
+}
+
