+++ src/lib/classes/automaton/system_edge.ts
@@ -1,0 +1,20 @@
+
+
+/**
+* Describes the edges of a system.
+* */
+export class SystemEdge {
+	/**
+	* The beginning of the edge
+	* */
+	parent: number;
+
+	/**
+	* The end of the edge
+	* */
+	child: number;
+	constructor(parent: number, child: number) {
+		this.parent = parent;
+		this.child = child;
+	}
+}
