+++ src/lib/classes/draw.ts
@@ -1,0 +1,3 @@
+export { Point } from './draw/point';
+export { Dimensions } from './draw/dimensions';
+
