+++ src/lib/classes/draw/dimensions.ts
@@ -1,0 +1,10 @@
+
+export class Dimensions {
+	width: number = 0;
+	height: number = 0;
+
+	constructor(width: number, height: number) {
+		this.width = width;
+		this.height = height;
+	}
+}
