+++ src/lib/classes/draw/point.ts
@@ -1,0 +1,11 @@
+
+export class Point {
+	x: number;
+	y: number;
+
+	constructor(x: number, y: number) {
+		this.x = x;
+		this.y = y;
+	}
+}
+
